Keep navigation code with root Composables
